ARM: EXYNOS4: PM Bugfix for Exynos4210-EVT1 not waking up from s2ram.
Exynos4210 EVT1 (s5pc210-evt1) has an issue of not waking up from
suspend-to-RAM WAKEUP_MASK register's 31'b is not cleared before
entering a suspend. This patch clears every undocumented bits of
WAKEUP_MASK before entering a suspend and enabled sleep/wake in the
specified chip.

Besides, EPLL and VPLL should be enabled when entering a suspend. The
status of EPLL/VPLL is resumed after wakeup anyway; they are saved and
restored.

Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>