ACPICA: Fixes unmerged (dis)assember divergences.
Assembler and Disassembler updates in the ACPICA were ignored in the past by
the ACPICA Linux release. This will introduce divergences between the Linux
and the ACPICA.
This patch ports the required updates of such utilities to Linux to reduce
the divergences.

This patch will not affect the generated vmlinux binary.
This will decrease 82 lines of 20121018 divergence.diff.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>